+++ src/core/geometries/bonds/shaders/fragment.ts
