+++ src/core/geometries/bonds/shaders/vertex.ts
