+++ src/core/geometries/particles/shaders/fragment.ts
